app/requirements.txt

ultralytics==8.3.40
fastapi==0.115.5
uvicorn[standard]==0.30.0
pillow==10.4.0
python-multipart==0.0.9
